Convert library to dynamically load and reload ngspice shared object
Ngspice seems to crash fatally upon certain errors such as parse errors. The library cannot be reloaded on-the-fly if it
is statically linked, which causes this wrapper library not to function afterwards. Now the DLL is loaded dynamically via
the operating system and can be reloaded upon errors.

KiCad does this too, likely because of the same issues. Some of the code from KiCad was adapted for this.